(MISC): Slightly re-structured layout;         Tidied up
