+++ src/main/kotlin/org/rust/cargo/runconfig/CargoCommandConfiguration.kt
@@ -12,7 +12,7 @@ import com.intellij.util.xmlb.XmlSerializer
 import org.jdom.Element
 import org.rust.cargo.runconfig.forms.CargoRunConfigurationEditorForm
 import org.rust.cargo.util.cargoProjectRoot
-import org.rust.cargo.projectSettings.toolchain
+import org.rust.cargo.project.settings.toolchain
 import org.rust.cargo.util.getModules
 
 class CargoCommandConfiguration(project: Project,
