--- src/main/kotlin/org/rust/cargo/toolchain/impl/CargoProjectState.kt
@@ -1,3 +1,0 @@
-package org.rust.cargo.toolchain.impl
-
-
