+++ src/main/kotlin/org/rust/ide/actions/RefreshCargoProjectAction.kt
@@ -1,4 +1,4 @@
-package org.rust.cargo
+package org.rust.ide.actions
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -6,8 +6,8 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.Project
 import com.intellij.util.containers.isNullOrEmpty
-import org.rust.cargo.projectSettings.toolchain
-import org.rust.cargo.toolchain.CargoMetadataService
+import org.rust.cargo.project.settings.toolchain
+import org.rust.cargo.project.watcher.CargoMetadataService
 import org.rust.cargo.toolchain.RustToolchain
 import org.rust.cargo.util.getModules
 import org.rust.cargo.util.getService
@@ -17,6 +17,11 @@ class RefreshCargoProjectAction : AnAction() {
         templatePresentation.text = "Refresh Cargo project"
         templatePresentation.description = "Update Cargo project information and download new dependencies"
     }
+
+    private val Project.modulesWithToolchains: List<Pair<Module, RustToolchain>>
+        get() = getModules().orEmpty().mapNotNull { module ->
+            module.toolchain?.let { Pair(module, it) }
+        }
 
     override fun update(e: AnActionEvent) {
         if (e.project?.modulesWithToolchains.isNullOrEmpty()) {
@@ -34,8 +39,3 @@ class RefreshCargoProjectAction : AnAction() {
         }
     }
 }
-
-private val Project.modulesWithToolchains: List<Pair<Module, RustToolchain>>
-    get() = getModules().orEmpty().mapNotNull { module ->
-        module.toolchain?.let { Pair(module, it) }
-    }
