+++ src/test/kotlin/org/rust/cargo/toolchain/CargoProjectResolveTestCase.kt
@@ -8,7 +8,8 @@ import com.intellij.psi.PsiReference
 import com.intellij.util.ui.UIUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.rust.cargo.RustWithToolchainTestCaseBase
-import org.rust.cargo.projectSettings.toolchain
+import org.rust.cargo.project.settings.toolchain
+import org.rust.cargo.project.watcher.CargoMetadataService
 import org.rust.cargo.util.cargoProject
 import org.rust.cargo.util.getService
